일단저장

+++ D3D/Client/pch.cpp
@@ -3,5 +3,5 @@
 
 unique_ptr<Core> core =make_unique<Core>();
 
-int32 WINDOW_WIDTH =1200;
-int32 WINDOW_HEIGHT =800;
+float WINDOW_WIDTH =1200.0f;
+float WINDOW_HEIGHT =800.0f;
